+++ src/main/java/io/github/torand/fastersql/dialect/MySqlDialect.java
@@ -16,6 +16,7 @@
 package io.github.torand.fastersql.dialect;
 
 import java.util.EnumSet;
+import java.util.List;
 import java.util.Optional;
 
 import static io.github.torand.fastersql.dialect.Capability.LIMIT_OFFSET;
@@ -51,6 +52,11 @@ public class MySqlDialect implements Dialect {
     @Override
     public String formatSubstringFunction(String operand, int startPos, int length) {
         return "substring(" + operand + ", " + startPos + ", " + length + ")";
+    }
+
+    @Override
+    public String formatConcatFunction(List<String> operands) {
+        return "concat(%s)".format(String.join(",", operands));
     }
 
     @Override