+++ src/test/java/io/github/torand/fastersql/statement/oracle/OracleInsertStatementTest.java
@@ -15,20 +15,164 @@
  */
 package io.github.torand.fastersql.statement.oracle;
 
+import io.github.torand.fastersql.domainmodel.ProductCategory;
+import io.github.torand.fastersql.domainmodel.PurchaseStatus;
+import io.github.torand.fastersql.statement.PreparableStatement;
 import org.junit.jupiter.api.Test;
 
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.time.Year;
+import java.util.Optional;
+import java.util.UUID;
+
+import static io.github.torand.fastersql.constant.Constants.$;
+import static io.github.torand.fastersql.datamodel.DataModel.PRODUCT;
+import static io.github.torand.fastersql.datamodel.DataModel.PURCHASE;
+import static io.github.torand.fastersql.function.singlerow.SingleRowFunctions.toChar;
+import static io.github.torand.fastersql.function.system.SystemFunctions.currentTimestamp;
+import static io.github.torand.fastersql.statement.Statements.insert;
+import static io.github.torand.fastersql.statement.Statements.select;
+import static io.github.torand.fastersql.util.RowValueMatchers.isNull;
+import static org.hamcrest.Matchers.is;
 
 public class OracleInsertStatementTest extends OracleTest {
 
     @Test
-    void getConnection() throws SQLException {
-        try (Connection conn = ds.getConnection(); Statement stmt = conn.createStatement()) {
-            ResultSet rs = stmt.executeQuery("select * from purchase");
-            rs.next();
-        }
+    void shouldRetrieveInsertedRow() {
+        final UUID id = UUID.randomUUID();
+
+        PreparableStatement stmt =
+            insert().into(PRODUCT)
+                .value(PRODUCT.ID, id)
+                .value(PRODUCT.NAME, "Apple MacBook Pro")
+                .value(PRODUCT.CATEGORY, ProductCategory.ELECTRONICS)
+                .value(PRODUCT.DESCRIPTION, "TBD")
+                .value(PRODUCT.PRICE, 24999)
+                .value(PRODUCT.STOCK_COUNT, 7);
+
+        statementTester()
+            .assertSql("""
+                insert into PRODUCT (ID, NAME, CATEGORY, DESCRIPTION, PRICE, STOCK_COUNT) \
+                values (?, ?, ?, ?, ?, ?)"""
+            )
+            .assertParams(id, "Apple MacBook Pro", ProductCategory.ELECTRONICS, "TBD", 24999, 7)
+            .assertAffectedRowCount(1)
+            .verify(stmt);
+
+        statementTester()
+            .assertRowCount(1)
+            .assertRow(1,
+                "PR_ID", is(id.toString()),
+                "PR_NAME", is("Apple MacBook Pro"))
+            .verify(
+                select(PRODUCT.ID, PRODUCT.NAME)
+                    .from(PRODUCT)
+                    .where(PRODUCT.ID.eq(id))
+            );
+    }
+
+    @Test
+    void shouldHandleExpressionAsValue() {
+        final UUID id = UUID.randomUUID();
+        final UUID customerId = UUID.fromString("9df03cd1-245f-4257-95e2-85cb5bd39ad8"); // Ola Nordmann
+        final Year year = Year.now();
+
+        PreparableStatement stmt =
+            insert().into(PURCHASE)
+                .value(PURCHASE.ID, id)
+                .value(PURCHASE.STATUS, PurchaseStatus.REGISTERED)
+                .value(PURCHASE.CUSTOMER_ID, customerId)
+                .value(PURCHASE.NOTES, toChar(currentTimestamp(), "yyyy"))
+                .value(PURCHASE.CREATED_TIME, currentTimestamp());
+
+        statementTester()
+            .assertSql("""
+                insert into PURCHASE (ID, STATUS, CUSTOMER_ID, NOTES, CREATED_TIME) \
+                values (?, ?, ?, to_char(current_timestamp, ?), current_timestamp)"""
+            )
+            .assertParams(id, PurchaseStatus.REGISTERED, customerId, "yyyy")
+            .assertAffectedRowCount(1)
+            .verify(stmt);
+
+        statementTester()
+            .assertRowCount(1)
+            .assertRow(1,
+                "PU_ID", is(id.toString()),
+                "PU_NOTES", is(year.toString()),
+                "PU_CREATED_YEAR", is(year.toString()))
+            .verify(
+                select(PURCHASE.ID, PURCHASE.NOTES, toChar(PURCHASE.CREATED_TIME, "yyyy").as("PU_CREATED_YEAR"))
+                    .from(PURCHASE)
+                    .where(PURCHASE.ID.eq(id))
+            );
+    }
+
+    @Test
+    void shouldHandleNullAsValue() {
+        final UUID id = UUID.randomUUID();
+
+        PreparableStatement stmt =
+            insert().into(PRODUCT)
+                .value(PRODUCT.ID, id)
+                .value(PRODUCT.NAME, "Apple MacBook Pro")
+                .value(PRODUCT.CATEGORY, ProductCategory.ELECTRONICS)
+                .value(PRODUCT.DESCRIPTION, $(null))
+                .value(PRODUCT.PRICE, 24999)
+                .value(PRODUCT.STOCK_COUNT, 7);
+
+        statementTester()
+            .assertSql("""
+                insert into PRODUCT (ID, NAME, CATEGORY, DESCRIPTION, PRICE, STOCK_COUNT) \
+                values (?, ?, ?, null, ?, ?)"""
+            )
+            .assertParams(id, "Apple MacBook Pro", ProductCategory.ELECTRONICS, 24999, 7)
+            .assertAffectedRowCount(1)
+            .verify(stmt);
+
+        statementTester()
+            .assertRowCount(1)
+            .assertRow(1,
+                "PR_ID", is(id.toString()),
+                "PR_NAME", is("Apple MacBook Pro"),
+                "PR_DESCRIPTION", isNull())
+            .verify(
+                select(PRODUCT.ID, PRODUCT.NAME, PRODUCT.DESCRIPTION)
+                    .from(PRODUCT)
+                    .where(PRODUCT.ID.eq(id))
+            );
+    }
+
+    @Test
+    void shouldHandleOptionalValues() {
+        final UUID id = UUID.randomUUID();
+
+        PreparableStatement stmt =
+            insert().into(PRODUCT)
+                .value(PRODUCT.ID, id)
+                .value(PRODUCT.NAME, "Apple MacBook Pro")
+                .value(PRODUCT.CATEGORY, ProductCategory.ELECTRONICS)
+                .value(PRODUCT.DESCRIPTION, Optional.empty())
+                .value(PRODUCT.PRICE, 24999)
+                .value(PRODUCT.STOCK_COUNT, 7);
+
+        statementTester()
+            .assertSql("""
+                insert into PRODUCT (ID, NAME, CATEGORY, PRICE, STOCK_COUNT) \
+                values (?, ?, ?, ?, ?)"""
+            )
+            .assertParams(id, "Apple MacBook Pro", ProductCategory.ELECTRONICS, 24999, 7)
+            .assertAffectedRowCount(1)
+            .verify(stmt);
+
+        statementTester()
+            .assertRowCount(1)
+            .assertRow(1,
+                "PR_ID", is(id.toString()),
+                "PR_NAME", is("Apple MacBook Pro"),
+                "PR_DESCRIPTION", isNull())
+            .verify(
+                select(PRODUCT.ID, PRODUCT.NAME, PRODUCT.DESCRIPTION)
+                    .from(PRODUCT)
+                    .where(PRODUCT.ID.eq(id))
+            );
     }
 }