+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- ...
+- Supporting the 'substring' and 'concat' scalar functions.
+- INSERT and UPDATE now support any expression as field values.
 
 ### Changed
 - ...